Move the error perspective to the main window
No need to have a separate file for a few lines of code
